crewai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
google-cloud-aiplatform==1.70.0
vertexai==1.70.0
